added tools blocks
