+++ blocks/quote/quote.js
