--- blocks/tools/sidekick/config.json
@@ -1,65 +1,0 @@
-{
-  "project": "Geeks Edge Deliver Services",
-  "plugins": [
-    {
-      "id": "asset-library",
-      "title": "AEM Assets",
-      "environments": [
-      "edit","dev"
-      ],
-      "url": "https://experience.adobe.com/solutions/CQ-assets-selectors/static-assets/resources/franklin/asset-selector.html?rail=false",
-      "isPalette": true,
-      "passConfig": true,
-      "paletteRect": "top: 50px; bottom: 10px; right: 10px; left: auto; width:400px; height: calc(100vh - 60px)"
-    },  
-    {
-      "id": "google-assist",
-      "title": "Google Assist",
-      "environments": [ "edit","dev", "preview", "live", "prod"],
-      "url": "https://www.google.com/search?q=edge+delivery+services",
-      "isPalette": true,
-      "passConfig":true,
-      "passReferrer": true,
-      "excludePaths":[
-        "**/fr/**"
-      ],
-      "includePaths": [
-        "**.docx**",
-        "**/fr/blocks/**"
-      ],
-      "paletteRect": "top: 50px; bottom: 10px; right: 10px; left: auto; width:400px; height: calc(100vh - 60px)"
-    },
-    {
-      "id": "tools",
-      "title": "Tools",
-      "isContainer": true
-    },
-    {
-      "containerId": "tools",
-      "id": "contactus",
-      "title": "Contact US",
-      "environments": [ "edit","dev", "preview", "live", "prod" ],
-      "url": "/tools/contactus/contactus.html?user=aem-geeks",
-      "isPalette": true,
-      "passConfig": false,
-      "includePaths": [ "**/:x**" ],
-      "paletteRect": "top: 50px; bottom: 10px; right: 10px; left: auto; width:500px; height: calc(100vh - 60px)"
-    },
-    {
-      "containerId": "tools",
-      "id": "localize",
-      "title": "Localize",
-      "environments": [ "edit","dev", "preview", "live", "prod" ],
-      "url": "https://www.microsoft.com/en-us/translator/business/translator-api/",
-      "passReferrer": true
-    },
-    {
-      "containerId": "tools",
-      "id": "youtube",
-      "title": "YouTube",
-      "environments": [ "edit","dev", "preview", "live", "prod" ],
-      "url": "https://www.youtube.com/@AEMGeeks",
-      "passReferrer": true
-    }
-  ]
-}
