--- tools/blocks/fragment/fragment.css
@@ -1,1 +1,0 @@
-/* stylelint-disable no-empty-source */
