+++ tools/sidekick/config.json
@@ -60,6 +60,14 @@
       "environments": [ "edit","dev", "preview", "live", "prod" ],
       "url": "https://www.youtube.com/@AEMGeeks",
       "passReferrer": true
+    },
+    {
+      "containerId": "library",
+      "id": "block library",
+      "title": "YouTube",
+      "environments": [ "edit","dev", "preview", "live", "prod" ],
+      "url": "/tools/sidekick/library.html",
+      "includePaths": ["**.docx**"]
     }
   ]
 }
